Migrating from SimpleDateFormatter to DateFormatter, refactoring CFW.Utils.Time to CFW.Time

+++ src/main/java/com/xresch/cfw/features/query/CFWQueryContext.java
@@ -59,7 +59,7 @@ public class CFWQueryContext{
 	/***********************************************************************************************
 	 * Get the timezoneoffsetMinutes time for this query.
 	 ***********************************************************************************************/
-	public long getTimezoneOffsetMinutes() {
+	public int getTimezoneOffsetMinutes() {
 		return timezoneOffsetMinutes;
 	}
